pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "corprecon"
version = "1.0.0"
description = "OSINT tool to map French corporate mandate networks"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Twux", email = "twux@proton.me"},
]

dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "typer[all]>=0.12",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "respx>=0.21",
]

[project.scripts]
corprecon = "corprecon.cli:app"

[tool.setuptools.packages.find]
include = ["corprecon*"]
